move source download-specific guard to a more appropriate module
